Move Color Ext functions to :taj
